added an optional command-line parameter to MatlabSignalProcessing: 	MatlabSignalProcessing.exe --MatlabWD=foo makes 'foo' the working directory of the matlab engine session.
The default value is "." which means the working directory of the MatlabSignalProcessing.exe
binary (usually the installation 'prog' directory). MatlabWD can be expressed as an absolute
path, or relative to this.

This means that default behaviour has changed slightly: matlab's wd was undefined before,
which meant that it usually operated in MATLABROOT (e.g. C:\MATLAB\R2007a\)



git-svn-id: http://www.bci2000.org/svn/trunk@1864 7ef59cdf-0624-0410-a474-e34ce222aafa